Move DecoratedComment to `visitor.rs`
